[APDV-82] Changing directory structure
