--- frontend/lib/Models/BackendResponse.dart
@@ -1,10 +1,0 @@
-class BackendResponse<T>{
-  final T data;
-  final String error;
-
-  BackendResponse(this.data, this.error);
-
-  BackendResponse.fromJson(Map json) :
-    data = json["data"],
-    error = json["error"];
-}
