+++ frontend/lib/Models/EnableField.dart
@@ -1,4 +1,15 @@
-import 'Unit.dart';
+
+
+class Unit {
+  int id;
+  String name;
+
+  Unit(this.id, this.name);
+
+  Unit.fromJson(Map map)
+      : id = map['id'] ?? -1,
+        name = map['name'] ?? "";
+}
 
 class EnableField {
   int id;
