+++ frontend/lib/Providers/CompareEndpointsProvider.dart
@@ -1,13 +1,13 @@
 import 'dart:math';
 
-import 'package:adpv_frontend/Common/Common.dart';
 import 'package:adpv_frontend/Repository/AbstractEndpointRepository.dart';
-import 'package:adpv_frontend/Repository/RestClient.dart';
+import 'package:adpv_frontend/Repository/RestEndpointRepository.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/cupertino.dart';
 
 import '../Models/Endpoint.dart';
 import '../Models/EndpointSummary.dart';
+import '../Utils/Consts.dart';
 
 class CompareEndpointsModel extends ChangeNotifier {
   Map<String, EndpointSummary> endpointSummaryMap = {};
@@ -15,7 +15,7 @@ class CompareEndpointsModel extends ChangeNotifier {
   Map<String, bool> selectedChips = {};
   List<String> commonFields = [];
   List<String> selectedEndpoints = [];
-  AbstractEndpointRepository repository = RestClient(Dio());
+  AbstractEndpointRepository repository = RestEnpointRepository(Dio());
 
   CompareEndpointsModel();
