+++ frontend/lib/Routing/Routing.dart
@@ -1,7 +1,9 @@
 // Consts for routing Endpoint List
 const String endpointListRoute = '/';
 const String endpointViewRoute = '/endpoint';
+
 // Consts for routing Profile
 const String profileRoute = '/';
+
 // Consts for routing Compare Endpoints
 const String compareEndpointsRoute = '/';
