+++ frontend/lib/Utils/Consts.dart
@@ -1,0 +1,4 @@
+const ignoreField = 'timestamp';
+const emptyField = 'Select Endpoint';
+const compareEndpointsViewAppBar =  'Compare charts';
+const endpointViewAppBar = 'Endpoint Details';
