+++ frontend/lib/Views/CompareEndpointsView.dart
@@ -1,13 +1,12 @@
-import 'package:adpv_frontend/Drawing/MultiDataChart.dart';
 import 'package:adpv_frontend/Models/EndpointSummary.dart';
 import 'package:adpv_frontend/Repository/AbstractEndpointRepository.dart';
+import 'package:adpv_frontend/Widgets/CompareEndpoints/MultiDataChart.dart';
 import 'package:flutter/material.dart';
 import 'package:multiselect/multiselect.dart';
 import 'package:provider/provider.dart';
-
-import '../Common/Common.dart';
 import '../Models/Endpoint.dart';
-import '../Providers/CompareEndpointsModel.dart';
+import '../Providers/CompareEndpointsProvider.dart';
+import '../Widgets/CommonWidgets.dart';
 
 class CompareChartsView extends StatefulWidget {
   const CompareChartsView({required this.repository, Key? key})
