+++ frontend/lib/main.dart
@@ -1,8 +1,8 @@
-import 'package:adpv_frontend/Repository/MockRepository.dart';
-import 'package:adpv_frontend/Repository/RestClient.dart';
+import 'package:adpv_frontend/Repository/RestEndpointRepository.dart';
 import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
+
 import 'App.dart';
 
 Future main() async {
@@ -12,8 +12,7 @@ Future main() async {
 
 class MyApp extends StatelessWidget {
   MyApp({Key? key}) : super(key: key);
-  final EndpointRepository endpointRepository = EndpointRepository.mock();
-  final RestClient restClient = RestClient(Dio());
+  final RestEnpointRepository restClient = RestEnpointRepository(Dio());
 
   @override
   Widget build(BuildContext context) => MaterialApp(
@@ -22,7 +21,6 @@ class MyApp extends StatelessWidget {
           primarySwatch: Colors.blue,
         ),
         home: App(
-          endpointRepository: endpointRepository,
           repository: restClient,
         ),
     );
